+++ test/controllers/request_controller_test.rb
@@ -16,27 +16,27 @@ class RequestControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should contribute tool and set location as public user" do
-    post contribute_tool_set_location_url, params: { community_tool: { url: "http://example.com" } }
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
+    post contribute_tool_set_location_url, params: { tool: { url: "http://example.com" } }
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
     assert_template :contribute_tool_about_me
     assert_response :success
   end
 
   test "should contribute tool and set location as regular user" do
     login(users(:valid))
-    assert_difference("CommunityTool.count") do
-      post contribute_tool_set_location_url, params: { community_tool: { url: "http://example.com" } }
+    assert_difference("Tool.count") do
+      post contribute_tool_set_location_url, params: { tool: { url: "http://example.com" } }
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
-    assert_equal users(:valid), assigns(:community_tool).user
-    assert_redirected_to contribute_tool_description_url(assigns(:community_tool))
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
+    assert_equal users(:valid), assigns(:tool).user
+    assert_redirected_to contribute_tool_description_url(assigns(:tool))
   end
 
   test "should not contribute tool with invalid url" do
-    post contribute_tool_set_location_url, params: { community_tool: { url: "not a url" } }
-    assert_not_nil assigns(:community_tool)
+    post contribute_tool_set_location_url, params: { tool: { url: "not a url" } }
+    assert_not_nil assigns(:tool)
     assert_template :contribute_tool_start
     assert_response :success
   end
@@ -44,41 +44,41 @@ class RequestControllerTest < ActionDispatch::IntegrationTest
   test "should contribute tool and register user as public user" do
     skip
     assert_difference("User.count") do
-      assert_difference("CommunityTool.count") do
-        post contribute_tool_register_user_url, params: { community_tool: { url: "http://example.com" }, user: { username: "newuser", email: "newuser@example.com" } }
+      assert_difference("Tool.count") do
+        post contribute_tool_register_user_url, params: { tool: { url: "http://example.com" }, user: { username: "newuser", email: "newuser@example.com" } }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
-    assert_equal "newuser@example.com", assigns(:community_tool).user.email
-    assert_redirected_to contribute_tool_description_url(assigns(:community_tool))
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
+    assert_equal "newuser@example.com", assigns(:tool).user.email
+    assert_redirected_to contribute_tool_description_url(assigns(:tool))
   end
 
   test "should contribute tool and assign logged in regular user from registration page" do
     login(users(:valid))
     assert_difference("User.count", 0) do
-      assert_difference("CommunityTool.count") do
-        post contribute_tool_register_user_url, params: { community_tool: { url: "http://example.com" }, user: { username: "newuser", email: "newuser@example.com" } }
+      assert_difference("Tool.count") do
+        post contribute_tool_register_user_url, params: { tool: { url: "http://example.com" }, user: { username: "newuser", email: "newuser@example.com" } }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
-    assert_equal users(:valid), assigns(:community_tool).user
-    assert_redirected_to contribute_tool_description_url(assigns(:community_tool))
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
+    assert_equal users(:valid), assigns(:tool).user
+    assert_redirected_to contribute_tool_description_url(assigns(:tool))
   end
 
   test "should not contribute tool and register user with existing email address" do
     skip
     assert_difference("User.count", 0) do
-      assert_difference("CommunityTool.count", 0) do
+      assert_difference("Tool.count", 0) do
         post contribute_tool_register_user_url, params: {
-          community_tool: { url: "http://example.com" },
+          tool: { url: "http://example.com" },
           user: { username: "duplicateaccount", email: "valid@example.com" }
         }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
     assert_not_nil assigns(:registration_errors)
     assert_template :contribute_tool_about_me
   end
@@ -86,57 +86,57 @@ class RequestControllerTest < ActionDispatch::IntegrationTest
   test "should contribute tool and sign in user" do
     users(:valid).update password: "password"
     assert_difference("User.count", 0) do
-      assert_difference("CommunityTool.count") do
+      assert_difference("Tool.count") do
         patch contribute_tool_sign_in_user_url, params: {
-          community_tool: { url: "http://example.com" },
+          tool: { url: "http://example.com" },
           email: "valid@example.com", password: "password"
         }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
-    assert_equal users(:valid), assigns(:community_tool).user
-    assert_redirected_to contribute_tool_description_url(assigns(:community_tool))
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
+    assert_equal users(:valid), assigns(:tool).user
+    assert_redirected_to contribute_tool_description_url(assigns(:tool))
   end
 
   test "should contribute tool and assign logged in regular user from sign in page" do
     login(users(:valid))
     assert_difference("User.count", 0) do
-      assert_difference("CommunityTool.count") do
+      assert_difference("Tool.count") do
         patch contribute_tool_sign_in_user_url, params: {
-          community_tool: { url: "http://example.com" }, email: "", password: ""
+          tool: { url: "http://example.com" }, email: "", password: ""
         }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
-    assert_equal users(:valid), assigns(:community_tool).user
-    assert_redirected_to contribute_tool_description_url(assigns(:community_tool))
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
+    assert_equal users(:valid), assigns(:tool).user
+    assert_redirected_to contribute_tool_description_url(assigns(:tool))
   end
 
   test "should not contribute tool and sign in user with invalid email and password" do
     assert_difference("User.count", 0) do
-      assert_difference("CommunityTool.count", 0) do
+      assert_difference("Tool.count", 0) do
         patch contribute_tool_sign_in_user_url, params: {
-          community_tool: { url: "http://example.com" },
+          tool: { url: "http://example.com" },
           email: "valid@example.com", password: ""
         }
       end
     end
-    assert_not_nil assigns(:community_tool)
-    assert_equal "http://example.com", assigns(:community_tool).url
+    assert_not_nil assigns(:tool)
+    assert_equal "http://example.com", assigns(:tool).url
     assert_not_nil assigns(:sign_in_errors)
     assert_template :contribute_tool_about_me
   end
 
   test "should get contribute tool description as regular user" do
     login(users(:valid))
-    get contribute_tool_description_url(community_tools(:draft))
+    get contribute_tool_description_url(tools(:draft))
     assert_response :success
   end
 
   test "should not get contribute tool description as public user" do
-    get contribute_tool_description_url(community_tools(:draft))
+    get contribute_tool_description_url(tools(:draft))
     assert_redirected_to new_user_session_url
   end
 
@@ -148,55 +148,55 @@ class RequestControllerTest < ActionDispatch::IntegrationTest
 
   test "should set description and publish tool as regular user" do
     login(users(:valid))
-    post contribute_tool_set_description_url(community_tools(:draft)), params: {
-      community_tool: { name: "Tool Name", description: "Tool Description" }
+    post contribute_tool_set_description_url(tools(:draft)), params: {
+      tool: { name: "Tool Name", description: "Tool Description" }
     }
-    assert_not_nil assigns(:community_tool)
-    assert_equal "Tool Name", assigns(:community_tool).name
-    assert_equal "Tool Description", assigns(:community_tool).description
-    assert_equal true, assigns(:community_tool).published?
-    assert_equal Time.zone.today, assigns(:community_tool).publish_date
-    assert_redirected_to tool_url(community_tools(:draft))
+    assert_not_nil assigns(:tool)
+    assert_equal "Tool Name", assigns(:tool).name
+    assert_equal "Tool Description", assigns(:tool).description
+    assert_equal true, assigns(:tool).published?
+    assert_equal Time.zone.today, assigns(:tool).publish_date
+    assert_redirected_to tool_url(tools(:draft))
   end
 
   test "should set description and save draft tool as regular user" do
     login(users(:valid))
-    post contribute_tool_set_description_url(community_tools(:draft)), params: {
-      community_tool: { name: "Tool Name - DRAFT", description: "" },
+    post contribute_tool_set_description_url(tools(:draft)), params: {
+      tool: { name: "Tool Name - DRAFT", description: "" },
       draft: "1"
     }
-    assert_not_nil assigns(:community_tool)
-    assert_equal "Tool Name - DRAFT", assigns(:community_tool).name
-    assert_equal "", assigns(:community_tool).description
-    assert_equal false, assigns(:community_tool).published?
-    assert_nil assigns(:community_tool).publish_date
-    assert_redirected_to tool_url(community_tools(:draft))
+    assert_not_nil assigns(:tool)
+    assert_equal "Tool Name - DRAFT", assigns(:tool).name
+    assert_equal "", assigns(:tool).description
+    assert_equal false, assigns(:tool).published?
+    assert_nil assigns(:tool).publish_date
+    assert_redirected_to tool_url(tools(:draft))
   end
 
   test "should not set description and submit tool as regular user with invalid id" do
     login(users(:valid))
     post contribute_tool_set_description_url(-1), params: {
-      community_tool: { name: "Tool Name", description: "Tool Description" }
+      tool: { name: "Tool Name", description: "Tool Description" }
     }
-    assert_nil assigns(:community_tool)
+    assert_nil assigns(:tool)
     assert_redirected_to dashboard_url
   end
 
   test "should not submit tool as regular user without description" do
     login(users(:valid))
-    post contribute_tool_set_description_url(community_tools(:draft)), params: {
-      community_tool: { name: "", description: "" }
+    post contribute_tool_set_description_url(tools(:draft)), params: {
+      tool: { name: "", description: "" }
     }
-    assert_not_nil assigns(:community_tool)
-    assert_equal ["can't be blank"], assigns(:community_tool).errors[:name]
-    assert_equal ["can't be blank"], assigns(:community_tool).errors[:description]
+    assert_not_nil assigns(:tool)
+    assert_equal ["can't be blank"], assigns(:tool).errors[:name]
+    assert_equal ["can't be blank"], assigns(:tool).errors[:description]
     assert_template :contribute_tool_description
     assert_response :success
   end
 
   test "should not set description and submit tool as public user" do
-    post contribute_tool_set_description_url(community_tools(:draft)), params: {
-      community_tool: { name: "Tool Name", description: "Tool Description" }
+    post contribute_tool_set_description_url(tools(:draft)), params: {
+      tool: { name: "Tool Name", description: "Tool Description" }
     }
     assert_redirected_to new_user_session_url
   end